Clean up deployment scripts [filtered branch]
__NB__ This is still something of a work in progress

__NB__ This branch should *never* be merged into develop/master

Why?

My understanding of this unit of work is that the deployment scripts in
this project are to be used as a guideline for building out a set of
generic deployment scripts; I.e., avoiding hard coding, using generic
docker images, etc.

Keeping this in view, I've created two branches

1. A branch with *only* the deployment scripts
2. A branch with the original repo, except that the deployment scripts
have been `git-subtree`'d in, for the purpose of testing branch 1.

The premise is that, since deployments of branch 2 will only work when
branch 1 is in the desired state (i.e., generally applicable to any
project), branch 1 can then be `git-subtree`'d (or less esoterically
"installed") in *to any existing project*.

Here is an example of how these scripts could be installed in any
project

```
git remote add wedding_api <wedding_api_remote>

git subtree add --prefix location/of/deploy/scripts --squash wedding_api deploy-scripts
```

__NB__ this branch could be pushed to it's own repo, and the references
to `wedding_api` in the above example would then be replaced by the new
repo name

__NB__ The changes in this commit are intended to maintain the existing
deployment behaviours, while altering the code to be more generic.